requirements.txt
pandas
matplotlib
seaborn
sqlalchemy
python-dotenv
adjustText
numpy